+++ app.config.ts
@@ -1,10 +1,9 @@
-import "dotenv/config";
 import { ExpoConfig, ConfigContext } from "@expo/config";
 
 const getIconForBuildEnv = () => {
   if (
     ["development", "staging", "preview"].includes(
-      process.env.EAS_BUILD_PROFILE
+      process.env.EAS_BUILD_PROFILE!
     )
   ) {
     return `./assets/icon-${process.env.EAS_BUILD_PROFILE}.png`;
@@ -23,27 +22,20 @@ type CustomConfig = Omit<ConfigContext, "config"> & {
 export default ({ config }: CustomConfig): ExpoConfig => ({
   ...config,
   icon: getIconForBuildEnv(),
-  hooks: {
-    postPublish: [
-      // Don't run sentry-expo/upload-sourcemaps if we're building on CI
-      !process.env.SENTRY_SKIP_RELEASE
-        ? {
-            file: "sentry-expo/upload-sourcemaps",
-            config: {
-              organization: process.env.SENTRY_ORG!,
-              project: process.env.SENTRY_PROJECT!,
-              authToken: process.env.SENTRY_AUTH_TOKEN!,
-              setCommits: true,
-            },
-          }
-        : {},
+  plugins: [
+    ...(config.plugins || []),
+    // Maybe we can remove the options and the plugin should inherit the values from the environment variables
+    [
+      "@sentry/react-native/expo",
+      {
+        organization: process.env.SENTRY_ORG,
+        project: process.env.SENTRY_PROJECT,
+      },
     ],
-  },
+  ],
   extra: {
     eas: {
       projectId: "bdcab203-d9ce-48aa-95db-938dbfe1be17",
     },
   },
 });
-
-//test