+++ src/hooks/auth-hooks.ts
@@ -1,15 +1,16 @@
+import * as Sentry from "@sentry/react-native";
 import { useQueryClient, useMutation } from "@tanstack/react-query";
 import * as Linking from "expo-linking";
 import * as WebBrowser from "expo-web-browser";
 import { WebBrowserRedirectResult } from "expo-web-browser";
+import { useEffect } from "react";
 import Toast from "react-native-toast-message";
-import * as Sentry from "sentry-expo";
 
-import { API_URL, http } from "../utils/http";
-import { APIAuthResponse, APIError } from "../utils/types";
 import { useRootStackNavigation } from "./navigation-hooks";
 import { useSecureStore } from "./use-secure-store";
 import { useUser } from "./user-hooks";
+import { API_URL, http } from "../utils/http";
+import { APIAuthResponse, APIError } from "../utils/types";
 
 type AuthPayload = {
   email: string;
@@ -31,7 +32,8 @@ export const useSignup = () => {
   const { setValueForKey } = useSecureStore();
   const queryClient = useQueryClient();
 
-  return useMutation(signup, {
+  return useMutation({
+    mutationFn: signup,
     onSuccess: async (data) => {
       await setValueForKey("accessToken", data.accessToken);
       await setValueForKey("refreshToken", data.refreshToken);
@@ -51,7 +53,8 @@ export const useLogin = () => {
   const { setValueForKey } = useSecureStore();
   const queryClient = useQueryClient();
 
-  return useMutation(login, {
+  return useMutation({
+    mutationFn: login,
     onSuccess: async (data) => {
       await setValueForKey("accessToken", data.accessToken);
       await setValueForKey("refreshToken", data.refreshToken);
@@ -60,7 +63,7 @@ export const useLogin = () => {
       await queryClient.refetchQueries(["user"]);
     },
     onError: (err: APIError) => {
-      console.error("Error during POST /login: ", err.response?.data.message);
+      console.error("Error during POST /login: ", err.response?.data?.message);
     },
   });
 };
@@ -69,7 +72,8 @@ export const useAnonymousLogin = () => {
   const { setValueForKey } = useSecureStore();
   const queryClient = useQueryClient();
 
-  return useMutation(anonymousLogin, {
+  return useMutation({
+    mutationFn: anonymousLogin,
     onSuccess: async (data) => {
       await setValueForKey("accessToken", data.accessToken);
       await setValueForKey("refreshToken", data.refreshToken);
@@ -78,7 +82,7 @@ export const useAnonymousLogin = () => {
     onError: (err: APIError) => {
       console.error(
         "Error signing in POST /anonymous",
-        err.response?.data.message
+        err.response?.data?.message
       );
     },
   });
@@ -108,11 +112,24 @@ export const useGitlabLogin = () => {
   const { setValueForKey } = useSecureStore();
   const queryClient = useQueryClient();
 
+  // https://docs.expo.dev/guides/authentication/#warming-the-browser
+  // Temporarily disabled due to the removal of OAuth support for Android
+  // useEffect(() => {
+  //   WebBrowser.warmUpAsync();
+
+  //   return () => {
+  //     WebBrowser.coolDownAsync();
+  //   };
+  // }, []);
+
   return async () => {
     try {
       const res = (await WebBrowser.openAuthSessionAsync(
         `${API_URL}/oauth/gitlab/authorize?state=${user.data?.id ?? ""}`,
-        Linking.createURL("/login/gitlab")
+        Linking.createURL("/login/gitlab"),
+        {
+          showInRecents: true,
+        }
       )) as WebBrowserRedirectResult;
 
       // Only "success" is a suppoterd type but this doesn't ensure that the
@@ -124,7 +141,7 @@ export const useGitlabLogin = () => {
 
       if (res.type !== "success") {
         console.error("Error response from OAuth: ", res);
-        Sentry.Native.captureException(
+        Sentry.captureException(
           new Error("Error response from OAuth: " + JSON.stringify(res))
         );
         return;