+++ src/screens/SignupScreen.tsx
@@ -1,6 +1,11 @@
 import { useFormik, FormikProvider } from "formik";
 import React from "react";
-import { ActivityIndicator, ScrollView, TouchableOpacity } from "react-native";
+import {
+  ActivityIndicator,
+  Platform,
+  ScrollView,
+  TouchableOpacity,
+} from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 
 import BackButton from "../components/BackButton";
@@ -129,7 +134,7 @@ const SignupScreen: React.FC<Props> = ({ navigation }) => {
                     <ErrorsList
                       errors={{
                         error:
-                          signup.error.response?.data.message ??
+                          signup.error.response?.data?.message ??
                           "Unknown Error",
                       }}
                     />
@@ -146,11 +151,13 @@ const SignupScreen: React.FC<Props> = ({ navigation }) => {
                   </Button>
                 </Box>
                 {/* Gitlab CTA */}
-                <Box marginTop="m">
-                  <Button onPress={loginWithGitlab} variant="outline">
-                    Continue with Gitlab
-                  </Button>
-                </Box>
+                {Platform.OS === "ios" && ( // TODO: remove this when we have a way to login with Gitlab on Android
+                  <Box marginTop="m">
+                    <Button onPress={loginWithGitlab} variant="outline">
+                      Continue with Gitlab
+                    </Button>
+                  </Box>
+                )}
               </>
             </FormikProvider>
             <Disclaimer />