+++ CHANGELOG.md
@@ -6,7 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
-- A toast is displayed when the internet connection is lost
+## [2.1.4] - 2024-06-17
+
+### Changes
+
+- Small internal changes
+
+## [2.1.3] - 2023-10-11
+
+### Changes
+
+- Minor internal stability changes
+
+## [2.1.2] - 2023-06-10
+
+### Fixed
+
+- Fixed GitLab login (Android)
 
 ## [2.1.1] - 2023-03-25
 
@@ -99,7 +115,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - First release on the stores
 
-[unreleased]: https://github.com/zaniluca/ping-4-gitlab/compare/v2.1.0...HEAD
+[unreleased]: https://github.com/zaniluca/ping-4-gitlab/compare/v2.1.4...HEAD
 [1.0.1]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v1.0.1
 [1.0.2]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v1.0.2
 [1.1.0]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v1.1.0
@@ -109,3 +125,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [2.0.0]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v2.0.0
 [2.1.0]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v2.1.0
 [2.1.1]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v2.1.1
+[2.1.2]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v2.1.2
+[2.1.3]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v2.1.3
+[2.1.4]: https://github.com/zaniluca/ping-4-gitlab/releases/tag/v2.1.4